BUGFIX:CloverDataFormatter - sometimes throwed NullPointerException
git-svn-id: https://svn.javlin.eu/svn/cloveretl.opensource/trunk@2626 a09ad3ba-1a0f-0410-b1b9-c67202f10d70

+++ cloveretl.engine/src/org/jetel/data/formatter/CloverDataFormatter.java
@@ -183,7 +183,9 @@ public class CloverDataFormatter implements Formatter {
 					}while (position == buffer.limit());
 					//clear up
 					idxTmpFile.delete();
-					idxTmpFile.getParentFile().delete();
+					if (idxTmpFile.getParentFile() != null) {
+						idxTmpFile.getParentFile().delete();
+					}					
 				}else{//out instanceof FileOutputStream
 					File idxInFile = new File(idxTmpFile.getCanonicalPath().substring(0,idxTmpFile.getCanonicalPath().lastIndexOf('.')));
 			    	//get last old index
